move migrations into erigon-db
